+++ sources/vec3/vec3_module.c
@@ -3,18 +3,18 @@
 /*                                                        :::      ::::::::   */
 /*   vec3_module.c                                      :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: ncarvalh <ncarvalh@student.42.fr>          +#+  +:+       +#+        */
+/*   By: crypto <crypto@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/07/07 23:08:51 by ncarvalh          #+#    #+#             */
-/*   Updated: 2023/07/07 23:08:54 by ncarvalh         ###   ########.fr       */
+/*   Updated: 2023/08/03 13:06:29 by crypto           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "miniRT.h"
 
-float	vec3_module(t_vec3 vec)
+double	vec3_module(t_vec3 vec)
 {
-	float	mod;
+	double	mod;
 
 	mod = sqrt(pow(vec.x, 2) + pow(vec.y, 2) + pow(vec.z, 2));
 	return (mod);
